refactor(server): extract /api/topic-ideas/* into a router module
Fifth route-group extraction. The cleanest yet: 5 contiguous handlers,
uniform requireAuth, only dependency is the shared pool, zero local-helper
blockers.

Moves all 5 topic-ideas handlers into src/server/routes/topic-ideas.js,
mounted `app.use('/api/topic-ideas', requireAuth, topicIdeasRouter)`
(router-level auth — all 5 authed).

Routes (paths byte-identical, guard-verified):
  GET    /api/topic-ideas/:brandProfileId
  POST   /api/topic-ideas              (bare -> router.post('/'))
  POST   /api/topic-ideas/bulk
  PATCH  /api/topic-ideas/:id
  DELETE /api/topic-ideas/:id

The bare POST /api/topic-ideas maps to router.post('/') (empty sub-path ->
'/'); the guard reconstructs it back to /api/topic-ideas via the mount.

Test plan:
- node --check server.js + module -> OK
- npm run lint (no-undef) -> clean (no surprises this time — pool-only)
- route-inventory guard -> 213 byte-identical (5 paths)
- vitest -> 93/93

Rollback: revert; verbatim move.
